refactor: дикомпозировал Input.tsx
