+++ Desktop/React +TS/my-power/src/components/types.ts
@@ -1,3 +1,4 @@
+import { useNavigate } from "react-router-dom";
 export type HabitFormProps = {
   habitInput: string;
   setHabitInput: React.Dispatch<React.SetStateAction<string>>;
@@ -30,14 +31,39 @@ export type DeleteParams = {
 export type ListItemsProps = {
   ways: WayItem[];
   setWays: React.Dispatch<React.SetStateAction<WayItem[]>>;
-  onDelete:(params :DeleteParams)=>void
-
-}
+  onDelete: (params: DeleteParams) => void;
+};
 export type UserHabitsProps = {
   habits: { id: string; label: string }[];
   handleDelete: (
     id: string,
-    setHabits: React.Dispatch<React.SetStateAction<{ id: string; label: string }[]>>
+    setHabits: React.Dispatch<
+      React.SetStateAction<{ id: string; label: string }[]>
+    >
   ) => void;
-  setHabits: React.Dispatch<React.SetStateAction<{ id: string; label: string }[]>>;
+  setHabits: React.Dispatch<
+    React.SetStateAction<{ id: string; label: string }[]>
+  >;
+};
+
+export type AuthData = {
+  name: string;
+  lastName: string;
+  email: string;
+};
+export type AuthFromProps= {
+  formData: AuthData;
+  setHasError: React.Dispatch<React.SetStateAction<FormErrors>>;
+  navigate: ReturnType<typeof useNavigate>
+};
+
+export type FormErrors = {
+  name: boolean;
+  lastName: boolean;
+  email: boolean;
+};
+export type HandleChangeProps = {
+  event: React.ChangeEvent<HTMLInputElement>;
+  setFormData: React.Dispatch<React.SetStateAction<AuthData>>;
+  setHasError: React.Dispatch<React.SetStateAction<FormErrors>>;
 };
