+++ Desktop/React +TS/my-power/src/pages/FirstFormPage.tsx
@@ -1,13 +1,13 @@
 import EffectHandle from "../components/fristInput";
-import  {  ToastContainer  } from 'react-toastify'
+import { ToastContainer } from "react-toastify";
 import { Header } from "../components/header";
 export default function FirstFormPage() {
   return (
     <main>
       <section>
-        <Header/>
+        <Header />
         <EffectHandle />
-         < ToastContainer  /> 
+        <ToastContainer />
       </section>
     </main>
   );
